Add TPM reset GPIO property and fix Measured Boot (#2)
* Disable devicetree measurement in Measured Boot

U-Boot applies dynamic fixups to the device tree during boot. These
runtime modifications cause the measured hash of the DTB to change,
leading to inconsistent PCR values.

This feature will remain disabled until a method to measure the
pre-fixup DTB or canonicalize the post-fixup DTB is implemented.

* Add reset-gpios property to TPM device node

Define the connection between the TPM RST pin and the SoC GPIO (RK_PC0).
This ensures the TPM is properly reset and initialized during the boot
sequence.